(fix) simplified filenames
