+++ scripts/v2_dman_v1_config.py
