+++ scripts/v2_dman_v1_multiple_pairs.py
