+++ scripts/v2_macd_bb_v1_config.py
